school

+++ project22/app/Http/Controllers/CompanyController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Company;
+use App\Models\CompanyType;
 use App\Http\Requests\StoreCompanyRequest;
 use App\Http\Requests\UpdateCompanyRequest;
 use Illuminate\Http\Request;
@@ -17,7 +18,8 @@ class CompanyController extends Controller
     public function index()
     {
 		$companies = Company::all();
-		return view('companies.index', ['companies'=>$companies]);
+		$companiesTypes = CompanyType::all();
+		return view('companies.index', ['companies'=>$companies],['companiesTypes' => $companiesTypes]);
     }
 
     /**
@@ -32,8 +34,8 @@ class CompanyController extends Controller
 		for($i=1; $i<=250;$i++){
 			$select_values[] = $i;
 		}
-		
-        return view('companies.create', ['select_values' => $select_values]);
+		$companiesTypes = CompanyType::all();
+        return view('companies.create', ['select_values' => $select_values],['companiesTypes' => $companiesTypes]);
     }
 
     /**
@@ -46,7 +48,7 @@ class CompanyController extends Controller
     {
         $company = new Company;
 		$company->company_name = $request->company_name;
-		$company->company_type = $request->company_type;
+		$company->company_type_id = $request->company_type_id;
 		$company->company_description = $request->company_description;
 		
 		$company->save();
@@ -61,7 +63,8 @@ class CompanyController extends Controller
      */
     public function show(Company $company)
     {
-        return view('companies.show', ['company'=>$company]);
+		$companiesTypes = CompanyType::all();
+        return view('companies.show', ['company'=>$company],['companiesTypes' => $companiesTypes]);
     }
 
     /**
@@ -72,8 +75,8 @@ class CompanyController extends Controller
      */
     public function edit(Company $company)
     {
-		
-       return view('companies.edit', ['company'=>$company]);
+		$companiesTypes = CompanyType::all();
+       return view('companies.edit', ['company'=>$company],['companiesTypes' => $companiesTypes]);
     }
 
     /**
@@ -86,7 +89,7 @@ class CompanyController extends Controller
     public function update(Request $request, Company $company)
     {
         $company->company_name = $request->company_name;
-		$company->company_type = $request->company_type;
+		$company->company_type_id = $request->company_type_id;
 		$company->company_description = $request->company_description;
 		
 		$company->save();
@@ -101,7 +104,11 @@ class CompanyController extends Controller
      */
     public function destroy(Company $company)
     {
+		$clients = $company->companyClients;
+		if(count($clients) != 0){
+			return redirect()->route('company.index')->with('error_message', 'Trinti negalima, kompanija turi klientų');
+		}
         $company->delete();
-		return redirect()->route('company.index');
+		return redirect()->route('company.index')->with('success_message', 'Sėkmingai ištrinta');
     }
 }
